get_mbx_size DVI$_DEVBUFSIZ requres long not short

+++ vms/vms_mbx_util.h
@@ -18,7 +18,7 @@ int simple_check_mbx(unsigned short channel);
 int map_fd_to_child(int fd, int pid);
 int read_mbx(int fd, char *buf, int size);
 int write_mbx_eof(int fd);
-unsigned short get_mbx_size(unsigned short channel);
+unsigned int get_mbx_size(unsigned short channel);
 int vms_isapipe(int fd);
 int vms_isapipe_by_name(char *name);
